feat: 裸金属支持纳管
